Removed uneccessary files
- Moved everything from msg/mail to msg/model
- Removed uneccessary files
