+++ src/msg/mod.rs
@@ -1,7 +1,5 @@
 pub mod cli;
-// pub mod model;
-// pub mod tpl;
+pub mod model;
 
-pub mod mail;
 pub mod attachment;
 pub mod envelope;
