--- src/msg/tpl/mod.rs
@@ -1,2 +1,0 @@
-pub mod cli;
-pub mod model;
